Refactor imports to respect the convention we introduced
