+++ simpleflow/executor.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 from ._decorators import deprecated
 
 if TYPE_CHECKING:
-    from simpleflow import Workflow
+    from simpleflow.workflow import Workflow
     from simpleflow.history import History
 
 __all__ = ["Executor"]
