+++ simpleflow/process/_named_mixin.py
