+++ simpleflow/swf/mapper/models/__init__.py
@@ -6,5 +6,5 @@
 from simpleflow.swf.mapper.models.activity import ActivityTask, ActivityType  # NOQA
 from simpleflow.swf.mapper.models.base import BaseModel  # NOQA
 from simpleflow.swf.mapper.models.domain import Domain  # NOQA
-from simpleflow.swf.mapper.models.history import History  # NOQA
+from simpleflow.swf.mapper.models.history.base import History  # NOQA
 from simpleflow.swf.mapper.models.workflow import WorkflowExecution, WorkflowType  # NOQA
