+++ simpleflow/swf/mapper/responses/__init__.py
@@ -1,1 +1,1 @@
-from .base import Response  # NOQA
+from ._base import Response  # NOQA
