+++ simpleflow/swf/mapper/responses/_base.py
