+++ simpleflow/swf/process/__init__.py
@@ -1,2 +1,2 @@
 from .poller import Poller  # NOQA  # isort:skip
-from .decider import Decider  # NOQA  # isort:skip
+from .decider.base import Decider  # NOQA  # isort:skip
