+++ simpleflow/swf/process/poller.py
@@ -12,7 +12,7 @@ from simpleflow.process import NamedMixin, with_state
 from simpleflow.swf.helpers import swf_identity
 
 if TYPE_CHECKING:
-    from simpleflow.swf.mapper.models import Domain
+    from simpleflow.swf.mapper.models.domain import Domain
 
 __all__ = ["Poller"]
