+++ simpleflow/swf/stats/__init__.py
@@ -1,2 +1,2 @@
-from .base import *  # NOQA  # isort:skip
+from .base import WorkflowStats  # NOQA  # isort:skip
 from . import pretty  # NOQA  # isort:skip
