+++ tests/data/workflows.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from simpleflow import Workflow
+from simpleflow.workflow import Workflow
 from simpleflow.constants import HOUR, MINUTE
