+++ tests/test_simpleflow/local/test_executor.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import unittest
 
-from simpleflow import Workflow
+from simpleflow.workflow import Workflow
 from simpleflow.activity import with_attributes
 from simpleflow.canvas import Chain, Group
 from simpleflow.constants import HOUR, MINUTE
