+++ tests/test_simpleflow/swf/process/test_worker.py
@@ -5,7 +5,8 @@ from collections import namedtuple
 from unittest.mock import patch
 
 from simpleflow.swf.process.worker.base import ActivityPoller, ActivityWorker
-from simpleflow.swf.mapper.models import ActivityTask, Domain
+from simpleflow.swf.mapper.models.activity import ActivityTask
+from simpleflow.swf.mapper.models.domain import Domain
 from tests.moto_compat import mock_swf
 
 FakeActivityType = namedtuple("FakeActivityType", ["name"])
